Test against various LTS versions
To help us figure out the bounds for our dependencies.

For no particularly good reason, I decided to just be compatible all the way
back to LTS 0 cause it was easy and required only minimal compromises, namely:

* no Data.List.uncons
* Eq is not assumed to be a superclass of Hashable
* hiding some Prelude exports and importing them directly from other modules
  Importing something like <*> more than once would cause warnings on newer
  versions. Warnings about Prelude not exporting <*> or whatever on older
  versions are fine.
* using cabal spec 1.18
  The only newer feature we used were common/import stanzas for -Wall.

On the most recent version of GHC and our dependencies, there's some warnings
left in the tests, but those can be addressed later. We're not gonna worry about
warnings from older versions.