Remove finite-typelits from test and bench
It's just too much of a pain to work with.

We need the Ix instance available in >= 0.2, but almost all LTS snapshots have
a 0.1 version without that instance.

Also, recent releases have somewhat strict version bounds on base (e.g <4.23),
while some older releases have more lax bounds (e.g. <5). This causes the
package to effectively be downgraded if you upgrade GHC and finite-typelits
hasn't been updated for that new GHC release yet, which is unexpected and very
confusing.

We could work around this somehow in a variety of ways, but it doesn't seem
worth it.

We'll keep the example in the documentation, though, as it seems useful and
illustrative. doctest is probably still broken because of it, but we'll figure
that out later.