Add bounds for transformers
Not really sure about the upper bound. Surely it'd work with newer releases
which don't even contain the module we need anymore? It's probably fine as is,
but we can always adjust things later if we're forcing some really old version
of transformers for no good reason in some cases.