Fix latex rendering
This was broken by enabling CPP, which turned \\ into a single literal \ and
got rid of the newline, and inserted an extra empty line to keep line numbers
aligned. This seems like the most straightforward fix instead of getting a
literal \\ to be emitted while also avoiding the extra newline which ends the
latex paragraph.